Rename utils to common
