+++ lib/src/core/common/analytics/analytics_reporter.dart
@@ -1,5 +1,5 @@
 import 'package:collection/collection.dart';
-import 'package:sizzle_starter/src/core/utils/analytics/firebase_analytics_reporter.dart';
+import 'package:sizzle_starter/src/core/common/analytics/firebase_analytics_reporter.dart';
 
 /// {@template analytics_reporter}
 /// Interface for reporting analytics events.
