+++ lib/src/core/common/layout/window_size.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:sizzle_starter/src/core/utils/extensions/context_extension.dart';
+import 'package:sizzle_starter/src/core/common/extensions/context_extension.dart';
 
 /// {@template window_size}
 /// Breakpoints for responsive design.
