+++ lib/src/feature/initialization/logic/app_runner.dart
@@ -4,11 +4,11 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sizzle_starter/src/core/constant/application_config.dart';
-import 'package:sizzle_starter/src/core/utils/app_bloc_observer.dart';
-import 'package:sizzle_starter/src/core/utils/bloc_transformer.dart';
-import 'package:sizzle_starter/src/core/utils/error_reporter/error_reporter.dart';
-import 'package:sizzle_starter/src/core/utils/logger/logger.dart';
-import 'package:sizzle_starter/src/core/utils/logger/printing_log_observer.dart';
+import 'package:sizzle_starter/src/core/common/app_bloc_observer.dart';
+import 'package:sizzle_starter/src/core/common/bloc_transformer.dart';
+import 'package:sizzle_starter/src/core/common/error_reporter/error_reporter.dart';
+import 'package:sizzle_starter/src/core/common/logger/logger.dart';
+import 'package:sizzle_starter/src/core/common/logger/printing_log_observer.dart';
 import 'package:sizzle_starter/src/feature/initialization/logic/composition_root.dart';
 import 'package:sizzle_starter/src/feature/initialization/widget/initialization_failed_app.dart';
 import 'package:sizzle_starter/src/feature/initialization/widget/root_context.dart';
