+++ lib/src/feature/initialization/logic/composition_root.dart
@@ -2,9 +2,9 @@ import 'package:clock/clock.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:sizzle_starter/src/core/constant/application_config.dart';
-import 'package:sizzle_starter/src/core/utils/error_reporter/error_reporter.dart';
-import 'package:sizzle_starter/src/core/utils/error_reporter/sentry_error_reporter.dart';
-import 'package:sizzle_starter/src/core/utils/logger/logger.dart';
+import 'package:sizzle_starter/src/core/common/error_reporter/error_reporter.dart';
+import 'package:sizzle_starter/src/core/common/error_reporter/sentry_error_reporter.dart';
+import 'package:sizzle_starter/src/core/common/logger/logger.dart';
 import 'package:sizzle_starter/src/feature/initialization/model/dependencies_container.dart';
 import 'package:sizzle_starter/src/feature/settings/bloc/app_settings_bloc.dart';
 import 'package:sizzle_starter/src/feature/settings/data/app_settings_datasource.dart';
