+++ lib/src/feature/settings/data/app_settings_datasource.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:sizzle_starter/src/core/utils/color_codec.dart';
-import 'package:sizzle_starter/src/core/utils/persisted_entry.dart';
+import 'package:sizzle_starter/src/core/common/color_codec.dart';
+import 'package:sizzle_starter/src/core/common/persisted_entry.dart';
 import 'package:sizzle_starter/src/feature/settings/data/theme_mode_codec.dart';
 import 'package:sizzle_starter/src/feature/settings/model/app_settings.dart';
 import 'package:sizzle_starter/src/feature/settings/model/app_theme.dart';
